requirements.txt
# Основные зависимости для обработки изображений и геоданных
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
rasterio>=1.2.0
geopandas>=0.9.0
shapely>=1.7.0
scikit-image>=0.18.0
matplotlib>=3.3.0
scipy>=1.7.0
fiona>=1.8.0
pyproj>=3.0.0
dask>=2023.0.0
distributed>=2023.0.0